backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
python-multipart==0.0.6
replicate==0.25.1
Pillow==10.2.0
redis==5.0.1
celery==5.3.6
aiogram==3.3.0
cryptography==42.0.1
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
boto3==1.34.23
